requirements.txt
# Core Streamlit and Web Framework
streamlit>=1.32.0
streamlit-option-menu>=0.3.6

# AI and Machine Learning
openai>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0

# Video and Media Processing
moviepy>=1.0.3
imageio-ffmpeg>=0.4.9
opencv-python>=4.8.0

# AI Video Generation (SVD)
diffusers>=0.21.0
transformers>=4.35.0
accelerate>=0.24.0
torch>=2.1.0
torchvision>=0.16.0
huggingface_hub>=0.17.0

# Utilities
python-dotenv>=1.0.0
imagehash>=4.3.1
yt-dlp>=2023.10.13

# Development and Quality
ruff>=0.1.0
mypy>=1.7.0

# Optional: For advanced features
# sentence-transformers>=2.2.2  # For text similarity
# chromadb>=0.4.0              # For vector database
# langchain>=0.0.350           # For LLM chains